Split out another example from network_unmixer.py
